move sql into migration
